Feature/composer mssql

+++ database/migrations/2021_08_16_235826_create_watchers_table.php
@@ -15,8 +15,8 @@ class CreateWatchersTable extends Migration
     {
         Schema::create('watchers', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id');
-            $table->integer('topic_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('topic_id');
             $table->foreign('user_id')->references('id')->on('users');
             $table->foreign('topic_id')->references('id')->on('topics');
         });
